指令佮函式庫拆開 丞宏
