+++ batch_align.py
